services/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.6
pydantic-settings>=2.0
firebase-admin>=6.4
python-dotenv>=1.0
bleak>=0.22.0
pyserial>=3.5
PyJWT>=2.8
bcrypt>=4.1
email-validator>=2.2
pytest>=8.4,<10